+++ src/main/java/com/jackalhan/ualr/domain/TypeSafeRawWorkload.java
@@ -25,16 +25,16 @@ public class TypeSafeRawWorkload implements Serializable, Cloneable {
     private String courseTitle;
     private String collCode;
     private String taStudent;
-    private int taSupport;
-    private int taEleventhDayCount;
-    private int taCeditHours;
-    private int taLectureHours;
+    private double taSupport;
+    private double taEleventhDayCount;
+    private double taCeditHours;
+    private double taLectureHours;
     private double iuMultipliertaLectureHours;
-    private int taLabHours;
+    private double taLabHours;
     private double iuMultipliertaLabHours;
     private double totalIus;
     private String otherInstructorsInTeam;
-    private int totalSsch;
+    private double totalSsch;
     private String chair;
     private String dean;
 
@@ -154,35 +154,35 @@ public class TypeSafeRawWorkload implements Serializable, Cloneable {
         this.taStudent = taStudent.trim();
     }
 
-    public int getTaSupport() {
+    public double getTaSupport() {
         return taSupport;
     }
 
-    public void setTaSupport(int taSupport) {
+    public void setTaSupport(double taSupport) {
         this.taSupport = taSupport;
     }
 
-    public int getTaEleventhDayCount() {
+    public double getTaEleventhDayCount() {
         return taEleventhDayCount;
     }
 
-    public void setTaEleventhDayCount(int taEleventhDayCount) {
+    public void setTaEleventhDayCount(double taEleventhDayCount) {
         this.taEleventhDayCount = taEleventhDayCount;
     }
 
-    public int getTaCeditHours() {
+    public double getTaCeditHours() {
         return taCeditHours;
     }
 
-    public void setTaCeditHours(int taCeditHours) {
+    public void setTaCeditHours(double taCeditHours) {
         this.taCeditHours = taCeditHours;
     }
 
-    public int getTaLectureHours() {
+    public double getTaLectureHours() {
         return taLectureHours;
     }
 
-    public void setTaLectureHours(int taLectureHours) {
+    public void setTaLectureHours(double taLectureHours) {
         this.taLectureHours = taLectureHours;
     }
 
@@ -194,11 +194,11 @@ public class TypeSafeRawWorkload implements Serializable, Cloneable {
         this.iuMultipliertaLectureHours = iuMultipliertaLectureHours;
     }
 
-    public int getTaLabHours() {
+    public double getTaLabHours() {
         return taLabHours;
     }
 
-    public void setTaLabHours(int taLabHours) {
+    public void setTaLabHours(double taLabHours) {
         this.taLabHours = taLabHours;
     }
 
@@ -218,11 +218,11 @@ public class TypeSafeRawWorkload implements Serializable, Cloneable {
         this.otherInstructorsInTeam = otherInstructorsInTeam.trim();
     }
 
-    public int getTotalSsch() {
+    public double getTotalSsch() {
         return totalSsch;
     }
 
-    public void setTotalSsch(int totalSsch) {
+    public void setTotalSsch(double totalSsch) {
         this.totalSsch = totalSsch;
     }
